requirements.txt
matplotlib
numpy
opencv_python
Pillow
scikit_learn
scipy
setuptools
scikit-image
tqdm
Pillow
torch 
torchvision 
torchaudio
scikit-learn
pandas
transformers
codecarbon
datasets
seaborn